requirements.txt
fastapi==0.111.0
uvicorn 
Pillow 
requests 
python-dotenv 
ffmpeg-python 
numpy 
openai 
validators 
dependency-injector
loguru
fastapi-key-auth
black
python-docx
anthropic==0.31.0
Levenshtein
av
google-generativeai
pandas
google-cloud-aiplatform 
vertexai
pdf2image
google-crc32c 
docstring-parser 
google-cloud-bigquery 
google-cloud-resource-manager 
packaging 
shapely 
google-generativeai
docx2pdf
pymongo
